pyproject.toml
[project]
name = "mle_take_home"
version = "0.1.0"
description = ""

readme = "README.md"
requires-python = "==3.11.7"

dependencies = [
    "torch (>=2.8.0,<3.0.0)",
    "torchvision (>=0.23.0,<0.24.0)",
    "lightning (>=2.5.3,<3.0.0)",
    "torchmetrics (>=1.8.1,<2.0.0)",
    "pillow (>=11.3.0,<12.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
